+++ Src/Orbiter/Pane.h
@@ -179,11 +179,11 @@ public:
 	double SetMFDRefreshIntervalMultiplier (int id, double multiplier);
 	// modify the refresh interval of an MFD
 
-	bool OpenMFD (int id, int type, std::ifstream *ifs = 0);
+	bool OpenMFD (size_t id, int type, std::ifstream *ifs = 0);
 	// open specified instrument as left/right MFD. Returns true if MFD mode has changed
 	// If scenario stream is provided, parameters are read from this stream
 
-	bool CloseMFD (int id);
+	bool CloseMFD (size_t id);
 	//  Deactivates the specified MFD. Returns false if id invalid or MFD not active
 
 	void MFDModeDisabled (int mode);
